# server/swagger.yaml
openapi: 3.0.0
info:
  title: Tech Notes API
  description: Node.js + MongoDB - API with email sign-up, verification, authentication and forgot password
  version: 1.0.0
servers:
  - url: https://technotes-api-main.herokuapp.com/
    description: Production server
paths:
...
  /api/posts:
    get:
      summary: Get a list of all post messages
      description: Restricted to admin users.
      operationId: getAllPosts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: An array of all post messsages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "60845f68f84d58892f0bd3bf"
                    title:
                      type: string
                      example: "Sample post title"
                    message:
                      type: string
                      example: "Sample post message"
                    name:
                      type: string
                      example: "Jason"
                    creator:
                      type: string
                      example: "1233456"
                    createdAt:
                      type: string
                      example: "2021-04-24T18:11:52.600Z"
                    updated:
                      type: string
                      example: "2021-04-24T18:11:52.600Z"
                    tags:
                      type: array
                      example:  ["tag1, foo, bar"]
                    likes:
                      type: array
                      example:  ["1233456"]
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
        summary: Create a new post. The `creator` field will be post owner user's `id` and the `name` field suppose to be post owner user's `userName`
        description: Restricted to admin users.
        operationId: createPost
        security:
          - bearerAuth: []
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                    example: "Sample post"
                  message:
                    type: string
                    example: "Sample post"
                  name:
                    type: string
                    example: "Jason (post owner 'userName')"
                  creator:
                    type: string
                    example: "1233456 (post owner 'id')"
                  tags:
                    type: array
                    example:  ["tag1, foo, bar"]
                required:
                  - name
                  - creator
                  - title
                  - message
        responses:
          "200":
            description: Post created successfully. The details of the new post are returned.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                      id:
                        type: string
                        example: "60845f68f84d58892f0bd3bf"
                      title:
                        type: string
                        example: "Sample post title"
                      message:
                        type: string
                        example: "Sample post message"
                      name:
                        type: string
                        example: "Jason"
                      creator:
                        type: string
                        example: "1233456"
                      createdAt:
                        type: string
                        example: "2021-04-24T18:11:52.600Z"
                      updated:
                        type: string
                        example: "2021-04-24T18:11:52.600Z"
                      tags:
                        type: array
                        example:  ["tag1, foo, bar"]
                      likes:
                        type: array
                        example:  []
          "401":
            $ref: "#/components/responses/UnauthorizedError"
  /api/posts/{id}:
    parameters:
      - in: path
        name: id
        description: Post id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a single post by id
      description: See a specific poss.
      operationId: getPostById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Details of the specified post
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "60845f68f84d58892f0bd3bf"
                  title:
                    type: string
                    example: "Sample post title"
                  message:
                    type: string
                    example: "Sample post message"
                  name:
                    type: string
                    example: "Jason"
                  creator:
                    type: string
                    example: "1233456"
                  createdAt:
                    type: string
                    example: "2021-04-24T18:11:52.600Z"
                  updated:
                    type: string
                    example: "2021-04-24T18:11:52.600Z"
                  tags:
                    type: array
                    example:  ["tag1, foo, bar"]
                  likes:
                    type: array
                    example:  ["1233456"]
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    put:
      summary: Update a post
      description: Admin users can update any post including role, regular users are restricted to their own post.
      operationId: updatePost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                  title:
                    type: string
                    example: "Sample post"
                  message:
                    type: string
                    example: "Sample post"
                  name:
                    type: string
                    example: "Jason"
                  creator:
                    type: string
                    example: "1233456"
                  tags:
                    type: array
                    example:  ["tag1, foo, bar"]
      responses:
        "200":
          description: Post updated successfully. The details of the updated post are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "60845f68f84d58892f0bd3bf"
                  title:
                    type: string
                    example: "Sample post title"
                  message:
                    type: string
                    example: "Sample post message"
                  name:
                    type: string
                    example: "Jason"
                  creator:
                    type: string
                    example: "1233456"
                  createdAt:
                    type: string
                    example: "2021-04-24T18:11:52.600Z"
                  updated:
                    type: string
                    example: "2021-04-24T18:11:52.600Z"
                  tags:
                    type: array
                    example:  ["tag1, foo, bar"]
                  likes:
                    type: array
                    example:  ["1233456"]
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    patch:
      summary: Patch to a post (only for LIKES for now!)
      description: Users can LIKES or DISLIKES any specific post.
      operationId: patchPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                  user_id:
                    type: string
                    example: "Current user id"
                  action:
                    type: string
                    example: "inc (insert) OR dec (remove)"
      responses:
        "200":
          description: Patch updated successfully. The details of the updated post are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "60845f68f84d58892f0bd3bf"
                  title:
                    type: string
                    example: "Sample post title"
                  message:
                    type: string
                    example: "Sample post message"
                  name:
                    type: string
                    example: "Jason"
                  creator:
                    type: string
                    example: "1233456"
                  createdAt:
                    type: string
                    example: "2021-04-24T18:11:52.600Z"
                  updated:
                    type: string
                    example: "2021-04-24T18:11:52.600Z"
                  tags:
                    type: array
                    example:  ["tag1, foo, bar"]
                  likes:
                    type: array
                    example:  ["1233456"]
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete a post
      description: Admin users can delete any post, regular users are restricted to their own posts.
      operationId: deletePost
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Post deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/posts?{searchBy}&{search}:
    parameters:
      - in: path
        name: searchBy
        description: Any post field (id, title, name, message, etc.)
        required: true
        example: "title"
        schema:
          type: string
      - in: path
        name: search
        description: Any value for searching value.
        required: true
        example: "Foo"
        schema:
          type: string
    get:
      summary: Search for all matches with a specific post field.
      description: Search all posts.
      operationId: getPostBySearch
      security:
        - bearerAuth: []
      responses:
        "200":
          description: An array of all post messsages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "60845f68f84d58892f0bd3bf"
                    title:
                      type: string
                      example: "Sample post title"
                    message:
                      type: string
                      example: "Sample post message"
                    name:
                      type: string
                      example: "Jason"
                    creator:
                      type: string
                      example: "1233456"
                    createdAt:
                      type: string
                      example: "2021-04-24T18:11:52.600Z"
                    updated:
                      type: string
                      example: "2021-04-24T18:11:52.600Z"
                    tags:
                      type: array
                      example:  ["tag1, foo, bar"]
                    likes:
                      type: array
                      example:  ["1233456"]
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /api/comments/post/{postmessage_id}:
    get:
      summary: Get whole comments of a specific post
      description: See whole comments.
      operationId: getCommentByPostMessageId
      security:
        - bearerAuth: []
      responses:
        "200":
          description: An array of the specified post message id comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "60845f68f84d58892f0bd3bf"
                  postmessage_id:
                    type: string
                    example: "60850f4ef86468349cf40b81"
                  message:
                    type: string
                    example: "Sample comment message"
                  creator:
                    type: string
                    example: "1233456"
                  createdAt:
                    type: string
                    example: "2021-04-24T18:11:52.600Z"
                  updated:
                    type: string
                    example: "2021-04-24T18:11:52.600Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"